chore: Refactor ReadHttpResponseMessage into a top-level class
The content of the class is the same in this commit
